api/pyproject.toml
[tool.poetry]
name = "points-list"
version = "0.1.0"
description = ""
authors = ["christopherwschultz <chris.william.schultz@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "points_list"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.6"
djangorestframework = "^3.14.0"
django-environ = "^0.9.0"
django-cors-headers = "^3.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
